Fix menu and other rendering for PCD8544
This is not ideal but now gets the job done. Ideally, the main should not be concerned about the refresh of the display. All should be handeled in the graphicdisplay.cpp. A simple solution would be to update periodically at a fixed FPS just like computers do. But ideally, the graphicdisplay.cpp would itself take care of the updates, but a wrapper for printing a single char would have to be created for use in the main in favor of using lcd_data() directly.

When I added the update to the end of lcd_data(), all rendering became quite slow, probably because of loops in lcd_string() or lcd_fixed_string().

Closes #6